visualizer/requirements.txt
pyserial>=3.5
numpy>=1.24
pyqtgraph>=0.13
PyQt6>=6.5
PyOpenGL>=3.1
